refactor rounds folder
